+++ src/firebase.js
@@ -8,6 +8,7 @@ import * as hrFunctions from './services/firebase/hrService';
 import * as clinicalFunctions from './services/firebase/clinicalService';
 import * as systemFunctions from './services/firebase/systemService';
 import * as maintenanceFunctions from './services/firebase/maintenanceService';
+import * as sesmtFunctions from './services/firebase/sesmtService';
 
 // Standard exports for the rest of the application
 export const authService = {
@@ -23,4 +24,5 @@ export const dbService = {
   ...clinicalFunctions,
   ...systemFunctions,
   ...maintenanceFunctions,
+  ...sesmtFunctions,
 };